Renaming base model file
