+++ src/database/tags/models.py
@@ -1,6 +1,6 @@
 from sqlalchemy import Column, String, Integer, ForeignKey
 from sqlalchemy.orm import relationship
-from src.database.base.models import UserOwnedModel
+from src.database.base.base_models import UserOwnedModel
 from src.database.db import db
